BDOG-3888: Added missed nullable `references` field for MongoDb format. Added test to deser from Mongo format to internal model.
